+++ History.md
@@ -3,8 +3,10 @@ Unpublished
 
   * allow sampling interval to be configured
   * pinMode returns this rather than undefined
+  * simplify overriding of digitalRead, digitalWrite and _tick
 
 0.0.0 / Feb 05 2017
 ===================
 
   * initial release
+